requirements.txt
openai
langchain
python-dotenv
streamlit
pillow
requests
gtts
